+++ lib/enumerize/set.rb
@@ -1,6 +1,7 @@
 module Enumerize
   class Set
     include Enumerable
+    include Predicatable
 
     attr_reader :values
 
@@ -52,6 +53,14 @@ module Enumerize
     end
 
     private
+
+    def define_query_method(value)
+      singleton_class.class_eval <<-RUBY, __FILE__, __LINE__ + 1
+        def #{value}?
+          include?("#{value}")
+        end
+      RUBY
+    end
 
     def mutate!
       @values = @obj.public_send("#{@attr.name}=", @values).values